changes per PR suggestions/requests
