+++ cbioportal_etl/scripts/cnv_merge_seg.py
@@ -5,20 +5,7 @@ import argparse
 import json
 from get_file_metadata_helper import get_file_metadata
 import concurrent.futures
-
-
-def resolve_config_paths(config, tool_dir):
-    """
-    Resolve paths dynamically based on assumptions:
-    - Paths starting with 'scripts/' or 'REFS/' are relative to the tool directory.
-    """
-    for key, value in config.items():
-        if isinstance(value, dict):
-            resolve_config_paths(value, tool_dir)
-        elif isinstance(value, str) and value.startswith(("REFS/", "scripts/", "external_scripts/")):
-            config[key] = os.path.abspath(os.path.join(tool_dir, value))
-
-    return config
+from cbioportal_etl.scripts.resolve_config_paths import resolve_config_paths
 
 
 def process_seg(cur_seg, new_seg, cbio_tum_id):
